requirements.txt
asgiref
django
django-extensions
djangorestframework
pillow
sqlparse
tzdata
